pyproject.toml
[project]
name = "trade-bot"
version = "0.1.0"
description = ""
authors = [
    {name = "Paulo Cesar Peixoto",email = "p.peixoyo@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "yfinance (>=0.2.61,<0.3.0)",
    "matplotlib (>=3.10.3,<4.0.0)",
    "schedule (>=1.2.2,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pandas-ta (>=0.3.14b0,<0.4.0)",
    "setuptools (>=80.9.0,<81.0.0)",
    "numpy (>=2.2.6,<3.0.0)",
    "pandas (>=2.2.3,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
